# swagger-export.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Get all users (Admin only)",
        "tags": [
          "👥 Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "👥 Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deactivateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deactivated successfully"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Deactivate user",
        "tags": [
          "👥 Users"
        ]
      }
    },
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "👥 Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Update own profile",
        "tags": [
          "👥 Users"
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "operationId": "UsersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User role updated successfully"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Update user role (Admin only)",
        "tags": [
          "👥 Users"
        ]
      }
    },
    "/users/leaderboard/xp": {
      "get": {
        "operationId": "UsersController_getLeaderboard",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard retrieved successfully"
          }
        },
        "security": [
          {
            "X-HTTP-TOKEN": []
          }
        ],
        "summary": "Get XP leaderboard",
        "tags": [
          "👥 Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "TechVerse Café API",
    "description": "Backend API for TechVerse Café MVP - A tech community platform with digital economy\n\n## 🚀 New Features in MVP\n- � **Digital Wallet & TechCoin System**: Earn, buy, and spend TechCoin\n- � **Challenges & Competitions**: Community challenges with rewards\n- �️ **Enhanced Projects System**: Task management with payments\n- 🌍 **Multi-language Support**: Localized content and UI\n- ☕ **Café Communities**: Topic-based discussion groups\n\n## Core Features\n- � JWT Authentication & Role-based Access Control\n- 📝 Content Management (Articles, Posts)\n- 🔖 Bookmarking System\n- 👥 User Profiles & Social Features\n- 🤖 AI Integration for content generation\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the `X-HTTP-TOKEN` header.\n\n## Role-Based Access Control\nThe platform uses a two-tier role system:\n- **Global Roles**: ADMIN, SUPERVISOR, EDITOR, MEMBER\n- **Café Roles**: BARISTA, THINKER, JOURNALIST, MEMBER\n\n## Getting Started\n1. Register a new user account\n2. Login to receive JWT token\n3. Use the token in `X-HTTP-TOKEN` header for authenticated requests\n4. Create your wallet and start earning TechCoin\n5. Join challenges and participate in projects\n\n## Download API Specification\n- JSON Format: [/api/swagger.json](/api/swagger.json)\n- YAML Format: [/api/swagger.yaml](/api/swagger.yaml)\n      ",
    "version": "2.0.0",
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "X-HTTP-TOKEN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-HTTP-TOKEN",
        "description": "JWT token for authentication. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "johndoe"
          },
          "bio": {
            "type": "string",
            "description": "User bio",
            "example": "Tech enthusiast and developer"
          },
          "avatar": {
            "type": "string",
            "description": "User avatar URL",
            "example": "https://example.com/avatar.jpg"
          },
          "countryId": {
            "type": "number",
            "description": "User country ID",
            "example": 1
          },
          "languageId": {
            "type": "number",
            "description": "User language ID",
            "example": 1
          }
        }
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "GUEST",
              "MEMBER",
              "THINKER",
              "JOURNALIST",
              "BARISTA"
            ],
            "example": "THINKER"
          }
        }
      },
...
    }
  }
}